requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic[email]>=2.0
pydantic-settings>=2.0
passlib[bcrypt]>=1.7.4
PyJWT>=2.8.0
